Extract out the cocoa breadcrumbs classes
